# apps/api-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for BoomChecker IoT system - device registration and management with JWT-based authentication",
        "title": "BoomChecker API",
        "termsOfService": "https://boomchecker.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@boomchecker.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/auth/request": {
            "post": {
                "description": "Request a JWT token for admin access. Token is sent via email and is valid for 24 hours. Rate limited to 1 request per 24 hours.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-auth"
                ],
                "summary": "Request admin authentication token",
                "parameters": [
                    {
                        "description": "Email address",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token request successful, email sent",
                        "schema": {
                            "$ref": "#/definitions/services.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized email",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "services.TokenRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@example.com"
                }
            }
        },
        "services.TokenResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "description": "UTC timestamp when token expires (RFC3339 format)",
                    "type": "string",
                    "example": "2025-11-13T12:00:00Z"
                },
                "message": {
                    "type": "string",
                    "example": "Admin token has been sent to your email"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token for admin authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token for node authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}